src/requirements.txt
django-ninja
Django~=4.2.4
uuid~=1.30
celery~=5.3.4
requests~=2.28.1
pytz~=2022.7.1
django-rest-swagger
drf_yasg
psycopg2-binary
pycron
